+++ gfx/drivers/gl1.c
@@ -630,7 +630,7 @@ static void gl1_raster_font_render_message(gl1_t *gl,
    for (;;)
    {
       const char *delim = strchr(msg, '\n');
-      size_t msg_len    = delim ? (delim - msg) : strlen(msg);
+      size_t msg_len    = delim ? (size_t)(delim - msg) : strlen(msg);
 
       /* Draw the line */
       gl1_raster_font_render_line(gl, font, glyph_q,
